Update 13-remove-conf.Rmd
Changed to use CPM normalization not SF because I realized the big effect of rounding might be due to introduction of lots of zeros due to dividing many counts of 1 by the SF. CPM normalization boosts the molecule counts higher so rounding shouldn't have such an impact. Plus looking at the Normalization page it looks like SF and UQ are back to doing nothing - not sure what has happened.